prototype/requirements.txt
# Data
pm4py>=2.7.0
pandas>=2.2.0
pyarrow>=15.0.0

# LLM & Agent
langgraph>=0.2.0
langchain-core>=0.3.0
langchain-anthropic>=0.3.0
langchain-openai>=0.3.0

# Data validation
pydantic>=2.7.0
pydantic-settings>=2.2.0

# Configuration
pyyaml>=6.0
python-dotenv>=1.0.0

# Dashboard
streamlit>=1.38.0
plotly>=5.22.0
graphviz>=0.20.0

# Testing
pytest>=8.0.0
